requirements.txt
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
boto3>=1.29.0
azure-identity>=1.15.0
azure-mgmt-authorization>=4.0.0
microsoft-graph-core>=1.0.0
msgraph-sdk>=1.2.0
requests>=2.31.0
pandas>=2.1.0
python-dotenv>=1.0.0
